Separate out File\CreateReadUpdateDeleteTest
