+++ tests/Integration/File/UpdateTest.php
@@ -1,0 +1,69 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SmartAssert\ApiClient\Tests\Integration\File;
+
+use SmartAssert\ApiClient\Exception\ClientException;
+use SmartAssert\ApiClient\Exception\Error\ErrorException;
+use SmartAssert\ApiClient\Exception\File\NotFoundException as FileNotFoundException;
+use SmartAssert\ServiceRequest\Error\BadRequestError;
+use SmartAssert\ServiceRequest\Error\BadRequestErrorInterface;
+use SmartAssert\ServiceRequest\Parameter\Parameter;
+use SmartAssert\ServiceRequest\Parameter\Requirements;
+
+class UpdateTest extends AbstractFileTestCase
+{
+    public function testUpdateUnauthorized(): void
+    {
+        $refreshableToken = self::$usersClient->createToken(self::USER1_EMAIL, self::USER1_PASSWORD);
+        $apiKey = self::$usersClient->getApiKey($refreshableToken->token);
+        $source = self::$fileSourceClient->create($apiKey->key, md5((string) rand()));
+
+        $filename = md5((string) rand()) . '.yaml';
+        self::$fileClient->create($apiKey->key, $source->getId(), $filename, md5((string) rand()));
+
+        $exception = null;
+
+        try {
+            self::$fileClient->update(md5((string) rand()), $source->getId(), $filename, md5((string) rand()));
+        } catch (ClientException $exception) {
+        }
+
+        self::assertInstanceOf(ClientException::class, $exception);
+
+        $fileNotFoundException = $exception->getInnerException();
+        self::assertInstanceOf(FileNotFoundException::class, $fileNotFoundException);
+    }
+
+    public function testUpdateEmptyFilename(): void
+    {
+        $refreshableToken = self::$usersClient->createToken(self::USER1_EMAIL, self::USER1_PASSWORD);
+        $apiKey = self::$usersClient->getApiKey($refreshableToken->token);
+        $source = self::$fileSourceClient->create($apiKey->key, md5((string) rand()));
+
+        $exception = null;
+
+        try {
+            self::$fileClient->update($apiKey->key, $source->getId(), '', md5((string) rand()));
+        } catch (ClientException $exception) {
+        }
+
+        self::assertInstanceOf(ClientException::class, $exception);
+
+        $errorException = $exception->getInnerException();
+        self::assertInstanceOf(ErrorException::class, $errorException);
+
+        $error = $errorException->getError();
+        self::assertInstanceOf(BadRequestErrorInterface::class, $error);
+
+        self::assertEquals(
+            new BadRequestError(
+                (new Parameter('filename', ''))
+                    ->withRequirements(new Requirements('yaml_filename')),
+                'invalid'
+            ),
+            $error
+        );
+    }
+}
